Moving towards passing tests. Backwards support of ghidra 11.3.2.
